Add Ui param to Storage creation
The previous commit changed the Storage constructor,
but the creation of the Storage object was not updated
to reflect that.

Improve logic flow of Commands and Parser

The bye and list commands have been added to the list
of Commands, and are no longer in their own separate
if else branch. Furthermore, the partialMatch method has
been removed in favour of simply checking in tryCommand,
since it is unnecessary.